third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/misc/client-hints-accept-parent-expected.txt

CONSOLE MESSAGE: line 7: success
This is a testharness.js-based test.
PASS Verify that hints were not sent on iframe subresources. 
Harness: the test ran to completion.
